Cambios Ajsutes de inventario con la facturacion. Cambio de tipo de varias variables para no perder informacion de int a decimal y tambien en la base de datos, columna con nro de fact o de recibido en el historial de inventario. Borrar productos POS

+++ Clients/Client.Main/Models/ClientesModel.cs
@@ -11,7 +11,7 @@ namespace Client.Main.Models
 {
     public class ClientesModel : PersonModel
     {
-        public int puntos { get; set; }
+        public decimal puntos { get; set; }
         public string estado { get; set; }
 
     }
