+++ Clients/Client.Main/Models/RecibidoModel.cs
@@ -11,7 +11,7 @@ namespace Client.Main.Models
         public DateTime fechaRecibido { get; set; } = DateTime.Now.Date;
         public string nombreConductor { get; set; }
         public string placas { get; set; }
-        public int? numeroCanastillas { get; set; } = null;
+        public decimal? numeroCanastillas { get; set; } = null;
         public int? peso { get; set; } = null;
 
         public BindableCollection<ProductoModel> productosActualizados { get; set; } = new BindableCollection<ProductoModel>();
