+++ Clients/Client.Main/ViewModels/PedidoResultadoBusquedaViewModel.cs
@@ -64,7 +64,7 @@ namespace Client.Main.ViewModels
                 short a;
                 if (Int16.TryParse(value, out a))
                 {
-                    if (Int16.Parse(value) < DateTime.Now.Day)
+                    if (Int32.Parse(value) < DateTime.Now.Day)
                     {
                         _dia = value;
                         NotifyOfPropertyChange(() => Dia);
@@ -85,13 +85,12 @@ namespace Client.Main.ViewModels
                 short a;
                 if (Int16.TryParse(value, out a))
                 {
-                    if (Int16.Parse(value) <= DateTime.Now.Month)
+                    if (Int32.Parse(value) <= DateTime.Now.Month)
                     {
                         _mes = value;
                         NotifyOfPropertyChange(() => Mes);
                     }
                 }
-
             }
         }
         private string _año = DateTime.Today.Year.ToString();
@@ -106,7 +105,7 @@ namespace Client.Main.ViewModels
                 short a;
                 if (Int16.TryParse(value, out a))
                 {
-                    if (Int16.Parse(value) < DateTime.Now.Year)
+                    if (Int32.Parse(value) < DateTime.Now.Year)
                     {
                         _año = value;
                         NotifyOfPropertyChange(() => Año);
