+++ Server/ServerConsole/Models/ClientesModel.cs
@@ -6,7 +6,7 @@ namespace ServerConsole.Models
 {
     public class ClientesModel : PersonModel
     {
-        public int puntos { get; set; }
+        public decimal puntos { get; set; }
         public string estado { get; set; }
 
     }
