+++ Server/ServerConsole/Models/ComprasModel.cs
@@ -9,7 +9,7 @@ namespace ServerConsole.Models
     {
 
         public int? numeroCanastillas { get; set; } = null;
-        public int? peso { get; set; } = null;
+        public decimal? peso { get; set; } = null;
 
         /// <summary>
         /// Lista con los pedidos que componen el documento de compras.
