+++ Server/ServerConsole/Models/RecibidoModel.cs
@@ -12,7 +12,7 @@ namespace ServerConsole.Models
         public string nombreConductor { get; set; }
         public string placas { get; set; }
         public int? numeroCanastillas { get; set; } = null;
-        public int? peso { get; set; } = null;
+        public decimal? peso { get; set; } = null;
         public BindableCollection<ProductoModel> productosActualizados { get; set; } = new BindableCollection<ProductoModel>();
 
         public static implicit operator RecibidoModel(EnvioModel v)
